refactor(workflow-compiler): make Strict fail fast on schema errors; dedup helpers
- Strict mode now throws on schema-propagation errors (e.g. a Projection on
  a missing column) instead of silently launching a plan that fails at
  runtime; the pinning test flips to assert the throw
- add Workflow.fromCompilationResult so the execution-path callers share the
  context-mutation + physicalPlan.get unpacking instead of open-coding it
- hoist getStackTraceWithAllCauses into workflow-core's StackTraceUtils;
  amber's ErrorUtils delegates and the compiler drops its private copy
